Last upload before due date
Worked on getting the tasks to work properly, but was not able to get the leds to toggle. Have added a .txt to the upload as well. Will continue working on this until I get the leds to function properly, and will send in video proof when it is complete. Though it will be late.